// packages/web/src/generated/rust-types/SshHostStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SshHostStats = { cpuUsagePercent: number, memTotalBytes: number, memUsedBytes: number, memAvailableBytes: number, memUsagePercent: number, swapTotalBytes: number, swapUsedBytes: number, memBuffersBytes: number, memCachedBytes: number, };